+++ src/index.js
@@ -6,7 +6,6 @@ export * from './effects'
 export * from './reducer'
 export * from './selectors'
 export * from './actions'
+export * from './hooks'
 export { default as combineRjs } from './combineRjs'
 export { default as rjMiddleware } from './rjMiddleware'
-export { default as useRj } from './useRj'
-export { default as useRunRj } from './useRunRj'
